Create a unified signAndSendTx function
Removed the dedicated function for auto-id
